fix: align profiling duration clocks
Help-Impact: none

Help-Impact-Reason: Internal profiling timing now uses one virtual-clock precision for stop decisions and CSV evidence; player inputs, Help-visible workflows, labels, and gameplay behavior are unchanged.

+++ crates/bevy_app/src/plugins/startup/perf_scenario.rs
@@ -172,7 +172,7 @@ use output::{
 #[derive(Resource, Default)]
 pub(crate) struct PerfCapture {
     phase: PerfCapturePhase,
-    elapsed_secs: f32,
+    elapsed_secs: f64,
     frame_times_ms: Vec<f64>,
     fixture_wait_reported: bool,
     initial_checksum: Option<PerfScenarioChecksum>,
